Finished organising the file structure
