+++ src/containers/homepage/carousel/CarouselSlide.js
